refactor(app): move last containers
